+++ LICENSE
@@ -217,4 +217,4 @@ dev/tox.ini
 BSD 3-Clause
 ------------
 
-databricks/koala/dask/*.py
+databricks/koalas/dask/*.py
