+++ databricks/koalas/tests/test_parquet.py
@@ -21,7 +21,7 @@ import numpy as np
 import pandas as pd
 import pyspark
 
-from databricks.koala.testing.utils import ReusedSQLTestCase, TestUtils
+from databricks.koalas.testing.utils import ReusedSQLTestCase, TestUtils
 
 
 class ParquetTest(ReusedSQLTestCase, TestUtils):
